tsconfig 설정 변경: BookDetailv.vue - 경로 에러 해결
