+++ myBookReport/src/views/BookDetail1.vue
@@ -135,10 +135,15 @@
     </div>
   </section>
 </template>
+
 <script setup lang="ts">
 import { ref } from 'vue'
+import { useTabChange } from '@/views/utils/changeInformation'
 
-const isActiveTab = ref<number>(1)
+// const isActiveTab = ref<number>(1)
+// const tapselect = ref<string>('introduce')
+
+const { isActiveTab, tapselect, changeTabActive } = useTabChange()
 
 const props = defineProps<{
   bookName: string
@@ -146,16 +151,14 @@ const props = defineProps<{
   bookUrl: string
 }>()
 
-const tapselect = ref<string>('introduce')
-
-const tabIndexConverTitle = [
-  { index: 1, title: 'introduce' },
-  { index: 2, title: 'review' },
-  { index: 3, title: 'order' },
-  { index: 4, title: 'write' },
-  { index: 5, title: 'recommendation' },
-  { index: 6, title: 'reference' }
-]
+// const tabIndexConverTitle = [
+//   { index: 1, title: 'introduce' },
+//   { index: 2, title: 'review' },
+//   { index: 3, title: 'order' },
+//   { index: 4, title: 'write' },
+//   { index: 5, title: 'recommendation' },
+//   { index: 6, title: 'reference' }
+// ]
 
 const bookinfolists = [
   { label: '저자', content: '김윤미' },
@@ -175,11 +178,11 @@ const tabs = ref([
   { label: '자료실', isDisabled: false }
 ])
 
-const changeTabActive = (index: number) => {
-  console.log(index)
-  isActiveTab.value = index
-  tapselect.value = tabIndexConverTitle[index].title
-}
+// const changeTabActive = (index: number) => {
+//   console.log(index)
+//   isActiveTab.value = index
+//   tapselect.value = tabIndexConverTitle[index].title
+// }
 </script>
 
 <style scoped>
