+++ myBookReport/src/views/service/bookDetailInfo.ts
@@ -1,0 +1,10 @@
+const tabIndexConverTitle = [
+  { index: 1, title: 'introduce' },
+  { index: 2, title: 'review' },
+  { index: 3, title: 'order' },
+  { index: 4, title: 'write' },
+  { index: 5, title: 'recommendation' },
+  { index: 6, title: 'reference' }
+]
+
+export default tabIndexConverTitle
