+++ myBookReport/tsconfig.app.json
@@ -6,7 +6,7 @@
     "composite": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["admin/src/*"]
+      "@/*": ["src/*"]
     }
   }
 }
