+++ myBookReport/tsconfig.node.json
@@ -12,9 +12,9 @@
     "module": "ESNext",
     "moduleResolution": "Bundler",
     "types": ["node", "vue"],
-    "path": {
+    "baseUrl": ".",
+    "paths": {
       "@/*": ["admin/src/*"]
-    },
-    
-  },
+    }
+  }
 }
